ft: complete fluent dot env options
