+++ src/dotenv.net/DotEnv.cs
@@ -6,7 +6,7 @@ using System.Text;
 
 namespace dotenv.net
 {
-    public class DotEnv
+    public static class DotEnv
     {
         private const string DefaultEnvFileName = ".env";
 
@@ -26,70 +26,23 @@ namespace dotenv.net
             }
         }
 
-        private static ReadOnlySpan<KeyValuePair<string, string>> ReadAndParse(string envFilePath,
-            bool ignoreExceptions, Encoding encoding, bool trimValues)
-        {
-            var rawEnvRows = Reader.Read(envFilePath, ignoreExceptions, encoding);
-
-            if (rawEnvRows == ReadOnlySpan<string>.Empty)
-            {
-                return ReadOnlySpan<KeyValuePair<string, string>>.Empty;
-            }
-
-            return Parser.Parse(rawEnvRows, trimValues);
-        }
-
-        private static IDictionary<string, string> ReadAndReturn(string envFilePath,
-            bool ignoreExceptions, Encoding encoding, bool trimValues)
-        {
-            var response = new Dictionary<string, string>();
-            var envRows = ReadAndParse(envFilePath, ignoreExceptions, encoding, trimValues);
-            foreach (var envRow in envRows)
-            {
-                if (response.ContainsKey(envRow.Key))
-                {
-                    response[envRow.Key] = envRow.Value;
-                }
-                else
-                {
-                    response.Add(envRow.Key, envRow.Value);
-                }
-            }
-
-            return response;
-        }
-
-        private static void ReadAndWrite(string envFilePath,
-            bool ignoreExceptions, Encoding encoding, bool trimValues)
-        {
-            var envRows = ReadAndParse(envFilePath, ignoreExceptions, encoding, trimValues);
-            foreach (var envRow in envRows)
-            {
-                Environment.SetEnvironmentVariable(envRow.Key, envRow.Value);
-            }
-        }
+       
 
         /// <summary>
         /// Initialize the fluent configuration API
         /// </summary>
-        /// <param name="throwOnError">A value stating whether the application should throw an exception on unexpected data</param>
-        /// <param name="filePath">An optional env file path, if not provided it defaults to the one in the same folder as the output exe or dll</param>
-        /// <param name="encoding">The encoding with which the env file was created, It defaults to the platforms default</param>
-        /// <param name="trimValues">This determines whether not whitespace is trimmed from the values. It defaults to true</param>
-        /// <exception cref="FileNotFoundException">Thrown if the env file doesn't exist</exception>
-        public static void Config(bool throwOnError = true, string filePath = DefaultEnvFileName,
-            Encoding encoding = null, bool trimValues = true)
+        public static DotEnvOptions Config()
         {
-            ConfigRunner(throwOnError, filePath, encoding, trimValues);
+            return new DotEnvOptions();
         }
 
         /// <summary>
         /// Configure the environment variables from a .env file
         /// </summary>
         /// <param name="options">Options on how to load the env file</param>
+        [Obsolete]
         public static void Config(DotEnvOptions options)
         {
-            ConfigRunner(options.IgnoreExceptions, options.EnvFilePaths, options.Encoding, options.TrimValues);
         }
 
         /// <summary>
@@ -97,6 +50,7 @@ namespace dotenv.net
         /// </summary>
         /// <param name="levelsToSearch">The number of top-level directories to search; the default is 4 top-level directories.</param>
         /// <returns>States whether or not the operation succeeded</returns>
+        [Obsolete]
         public static bool AutoConfig(int levelsToSearch = 4)
         {
             var currentDirectory = new DirectoryInfo(AppContext.BaseDirectory);
@@ -121,28 +75,8 @@ namespace dotenv.net
         /// <param name="envFilePath">The path to the .env file to be read</param>
         /// <param name="encoding">The encoding that the env file was saved in</param>
         /// <param name="ignoreExceptions">Determines if an exception should be thrown or swallowed</param>
-        public static Dictionary<string, string> Read(string envFilePath = DefaultEnvFileName, Encoding encoding = null,
-            bool ignoreExceptions = true)
+        public static Dictionary<string, string> Read(DotEnvOptions options)
         {
-            ConfigRunner(ignoreExceptions, envFilePath, encoding, true);
-        }
-
-        /// <summary>
-        /// Read and return the values in the provided env files
-        /// </summary>
-        /// <param name="envFilePaths">The paths to the .env files to be read</param>
-        /// <param name="encoding">The encoding that the env file was saved in</param>
-        /// <param name="ignoreExceptions">Determines if an exception should be thrown or swallowed</param>
-        /// <returns>An enumerable of dictionaries representing the contents of each env file</returns>
-        public static IEnumerable<Dictionary<string, string>> Read(IEnumerable<string> envFilePaths = null, Encoding encoding = null,
-            bool ignoreExceptions = true)
-        {
-            envFilePaths ??= Enumerable.Empty<string>();
-
-            foreach (var envFilePath in envFilePaths)
-            {
-                ConfigRunner(ignoreExceptions, envFilePath, encoding, true);
-            }
         }
 
         /// <summary>
@@ -151,27 +85,9 @@ namespace dotenv.net
         /// <param name="envFilePath">The path to the .env file to be read</param>
         /// <param name="encoding">The encoding that the env file was saved in</param>
         /// <param name="ignoreExceptions">Determines if an exception should be thrown or swallowed</param>
-        public static void Load(string envFilePath = DefaultEnvFileName, Encoding encoding = null,
-            bool ignoreExceptions = true)
+        public static void Load(DotEnvOptions options)
         {
-            ConfigRunner(ignoreExceptions, envFilePath, encoding, true);
-        }
-
-        /// <summary>
-        /// Load the values in the provided env files into the environment variables
-        /// </summary>
-        /// <param name="envFilePaths">The paths to the .env files to be read</param>
-        /// <param name="encoding">The encoding that the env file was saved in</param>
-        /// <param name="ignoreExceptions">Determines if an exception should be thrown or swallowed</param>
-        public static void Load(IEnumerable<string> envFilePaths = null, Encoding encoding = null,
-            bool ignoreExceptions = true)
-        {
-            envFilePaths ??= Enumerable.Empty<string>();
-
-            foreach (var envFilePath in envFilePaths)
-            {
-                ConfigRunner(ignoreExceptions, envFilePath, encoding, true);
-            }
+            
         }
     }
 }
